optimize logo
I like to optimize images by first resizing their dimensions to something reasonable, then optimizing them with ImageOptim: https://imageoptim.com/mac

This graphic is only displaying at 45px high, so I multiplied that by 4 to get 180px high, so that even the fanciest devices will see it in maximum resolution. 1000px is unnecessary.

I find that most PNGs are ok with a lossy 60% optimization, but I like tweaking and testing by hand.

In the long run, the text in this PNG should be actual text. We can color the "RI" using a CSS class and a span.